+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageClientImpl.java
@@ -1681,6 +1681,7 @@ public class GoogleCloudStorageClientImpl extends ForwardingGoogleCloudStorage
                     destination.hasGenerationId()
                         ? destination.getGenerationId()
                         : getWriteGeneration(destination, true)))
+            .setDeleteSourceObjects(options.isDeleteSourceObjects())
             .build();
 
     Blob composedBlob;